api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "prometheus-client>=0.25.0",
    "python-dotenv>=1.2.2",
    "redisvl>=0.16.0",
    "sentence-transformers>=5.4.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
    "pre-commit-hooks>=6.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
